Confirmation dialog - Confirmation on request and trip removal from account

+++ src/app/shared/post/post.component.ts
@@ -9,11 +9,12 @@ import { environment } from '@env/environment';
 import { Luggage } from '@models/luggage.model';
 import { Router } from '@angular/router';
 import { PostService } from '@core/post.service';
-import { MatSnackBar } from '@angular/material';
+import { MatSnackBar, MatDialog } from '@angular/material';
 import { Item } from '@models/item.model';
 import { ServerResponse } from '@models/app/server-response.model';
 import { HttpErrorResponse } from '@angular/common/http';
 import { PexelsService } from '@core/pexels.service';
+import { ConfirmComponent } from '@core/dialogs/confirm/confirm.component';
 @Component({
   selector: 'app-post',
   templateUrl: './post.component.html',
@@ -39,6 +40,7 @@ export class PostComponent implements OnInit, OnChanges {
   constructor(
     private router: Router,
     private postService: PostService,
+    private dialog: MatDialog,
     private snack: MatSnackBar,
     private pexels: PexelsService,
   ) { }
@@ -100,19 +102,33 @@ export class PostComponent implements OnInit, OnChanges {
   }
 
   removePost() {
-    const removalService = this.isTrip
-    ? this.postService.removeTrip(this.post as Trip)
-    : this.postService.removeRequest(this.post as Request);
-    removalService.subscribe((serverResponse) => {
-      if (serverResponse.status) {
-        this.snack.open(`${this.isTrip ? 'Le trajet' : 'L\'annonce'} a bien été supprimé${this.isTrip ? '' : 'e'}`,
-                        'OK', {duration: 3000});
-        this.remove.emit();
-      } else {
-        const snackRef = this.snack.open('Erreur lors de la suppression', 'Réessayer', {duration: 3000});
-        snackRef.onAction().subscribe(() => this.removePost());
+    const dialogRef = this.dialog.open(ConfirmComponent, {
+      height: '40vh',
+      width: '75vw',
+      data: {
+        title: `Supprimer un${this.isTrip ? ' trajet' : 'e annonce'}`,
+        message: `Cette action est irreversible. Voulez-vous continuez ?`,
+        action: 'Supprimer',
+        actionStyle: 'btn-danger',
       }
-    }, (error: HttpErrorResponse) => this.serverError(error.status));
+    });
+    dialogRef.afterClosed().subscribe((action: boolean) => {
+      if (action) {
+        const removalService = this.isTrip
+        ? this.postService.removeTrip(this.post as Trip)
+        : this.postService.removeRequest(this.post as Request);
+        removalService.subscribe((serverResponse) => {
+          if (serverResponse.status) {
+            this.snack.open(`${this.isTrip ? 'Le trajet' : 'L\'annonce'} a bien été supprimé${this.isTrip ? '' : 'e'}`,
+                            'OK', {duration: 3000});
+            this.remove.emit();
+          } else {
+            const snackRef = this.snack.open('Erreur lors de la suppression', 'Réessayer', {duration: 3000});
+            snackRef.onAction().subscribe(() => this.removePost());
+          }
+        }, (error: HttpErrorResponse) => this.serverError(error.status));
+      }
+    });
   }
 
   toPost() {
